+++ bin/heatpack.js
@@ -1,7 +1,9 @@
 #!/usr/bin/env node
 
+var fs = require('fs')
 var path = require('path')
 var parseArgs  = require('minimist')
+var tmp = require('tmp')
 
 var pkg = require('../package.json')
 var config = require('../webpack.config')
@@ -10,7 +12,8 @@ var server = require('../server')
 var args = parseArgs(process.argv.slice(2), {
   alias: {
     h: 'help',
-    v: 'version'
+    v: 'version',
+    f: 'force'
   }
 })
 
@@ -20,9 +23,29 @@ if (args.version) {
 }
 
 if (args.help || args._.length === 0) {
-  console.log('Usage: heatpack script.js')
+  console.log('Usage: heatpack [options] script.js')
+  console.log('')
+  console.log('Options:')
+  console.log('  -f, --force  Force heatpack to run the given script (disables React.render check)')
   process.exit(0)
 }
 
-config.entry.push(path.join(process.cwd(), args._[0]))
+var entryPath = path.join(process.cwd(), args._[0])
+
+if (!args.force) {
+  var code = fs.readFileSync(entryPath).toString()
+  if (code.indexOf('React.render') === -1) {
+    console.log("Couldn't find React.render in " + entryPath + " - assuming it exports a React component.")
+    var tempFile = tmp.fileSync({prefix: 'heatpack-', postfix: '.js'})
+    var componentPath = entryPath.replace(/\\/g, '\\\\')
+    fs.writeSync(tempFile.fd, [
+      "var React = require('react')",
+      "var Component = require('" + componentPath + "')",
+      "React.render(<Component/>, document.querySelector('#app'))"
+    ].join('\n'))
+    entryPath = tempFile.name
+  }
+}
+
+config.entry.push(entryPath)
 server(config)