+++ bin/heatpack.js
@@ -10,9 +10,14 @@ var server = require('../server')
 
 var args = parseArgs(process.argv.slice(2), {
   alias: {
+    f: 'force',
     h: 'help',
-    v: 'version',
-    f: 'force'
+    p: 'port',
+    v: 'version'
+  },
+  boolean: ['help', 'force', 'version'],
+  default: {
+    port: 3000
   }
 })
 
@@ -20,27 +25,30 @@ if (args.version) {
   console.log('v' + pkg.version)
   process.exit(0)
 }
-
 if (args.help || args._.length === 0) {
   console.log('Usage: heatpack [options] script.js')
   console.log('')
   console.log('Options:')
-  console.log('  -f, --force  Force heatpack to run the given script (disables React.render check)')
+  console.log("  -v, --version print heatpack's version")
+  console.log('  -p, --port    port to run the webpack dev server on [default: 3000]')
+  console.log('  -f, --force   force heatpack to run the given script (disable React.render check)')
   process.exit(0)
 }
 
-var entryPath = path.join(process.cwd(), args._[0])
+var options = {
+  alias: {},
+  entry: path.join(process.cwd(), args._[0]),
+  port: args.port
+}
 
 if (!args.force) {
-  var code = fs.readFileSync(entryPath).toString()
+  var code = fs.readFileSync(options.entry).toString()
   if (code.indexOf('React.render') === -1) {
-    console.log("Couldn't find React.render in " + entryPath + " - assuming it exports a React component.")
-    config.resolve.alias = {
-      'theydoitonpurposelynn': entryPath
-    }
-    entryPath = path.join(__dirname, '../dummy.js')
+    console.log("Couldn't find React.render in " + options.entry +
+                ' - assuming it exports a React component.')
+    options.alias['theydoitonpurposelynn'] = options.entry
+    options.entry = path.join(__dirname, '../dummy.js')
   }
 }
 
-config.entry.push(entryPath)
-server(config)
+server(config(options), options)