+++ webpack.config.js
@@ -1,5 +1,6 @@
 'use strict'
 
+var fs = require('fs')
 var path = require('path')
 var webpack = require('webpack')
 
@@ -7,8 +8,25 @@ var HEATPACK_MODULES = path.join(__dirname, 'node_modules')
 var NODE_MODULES_RE = /node_modules/
 var DUMMY_ENTRY_RE = /react-heatpack[\\/]dummy.js$/
 
-// We need to special-case loading the heatpack dummy entry module as it will
-// be under global node_modules.
+/**
+ * Tries to find a node_modules directory which will be resolved for requires
+ * from the working directory.
+ */
+function findWorkingDirNodeModules() {
+  var parts = process.cwd().split(path.sep)
+  while (parts.length > 0) {
+    var target = path.join(parts.join(path.sep), 'node_modules')
+    if (fs.existsSync(target)) {
+      return target
+    }
+    parts.pop()
+  }
+}
+
+/**
+ * We need to special-case exclusion to allow the heatpack dummy entry module to
+ * be processed by loaders, as it will be under global node_modules.
+ */
 function excludeJS(absPath) {
   if (DUMMY_ENTRY_RE.test(absPath)) {
     return false
@@ -37,14 +55,18 @@ module.exports = function config(options) {
       })
     ],
     resolve: {
-      alias: options.alias,
+      // If there's a node_modules in scope from where the user ran heatpack, we
+      // want to pick up React from it first, so prepend it to the list of
+      // directories to resolve modules from.
+      root: findWorkingDirNodeModules(),
       extensions: ['', '.js', '.jsx', '.cjsx', '.coffee'],
       // Resolve webpack dev server entry modules from heatpack's dependencies
-      fallback: HEATPACK_MODULES
+      fallback: HEATPACK_MODULES,
+      alias: options.alias
     },
     resolveLoader: {
-      // Resolve loaders from heatpack's dependencies
-      modulesDirectories: [HEATPACK_MODULES],
+      // Always resolve loaders from heatpack's own dependencies
+      root: HEATPACK_MODULES
     },
     module: {
       loaders: [