+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+**Changed:** `require('React')` now resolves to the `node_modules` in scope for the working directory, if there is one. It previously always resolved to the version from heatpack's own dependencies.
+
 **Added:** `-p`/`--port` option to specify the port to run the webpack dev server on. Defaults to 3000.
 
 ## v1.2.0 - 2015-07-04 01:00