+++ CHANGELOG.md
@@ -1,4 +1,8 @@
+## v1.3.0 - 2015-07-11 17:08
+
 **Changed:** `require('react')` now resolves to the `node_modules` dir in scope for the working directory, if there is one. It previously always resolved to the version from heatpack's own dependencies.
+
+**Changed:** enabled coloured output in webpack logging.
 
 **Added:** `-p`/`--port` option to specify the port to run the webpack dev server on. Defaults to 3000.